requirements.txt
matplotlib==3.3.4
numpy==1.19.5
pandas==1.1.5
pandas-plink==2.2.4
scikit-allel==1.3.5
scikit-learn==0.24.2
scipy==1.5.4
